refactor(plugin): :fire: refactoring, code cleanup and integration testing
